Add files via upload
I am pleased to share that I completed a project named as " Vrinda Store Sales" on Excel. Basically "Vrinda Store" is an online sales store dealing in clothing of both men and women. The channels through they are selling their product like Amazon, Meesho, Myntra, Ajio, Flipkart etc. 

Vrinda Store wants to know that which month has maximum sales, between men and women who are more likely to buy their products, in which states, order status and which channel has maximum contribution and the minimum as well. 

STEP-1 - DATA CLEANING (in which I made the data as uniform)

STEP-2 - DATA PROCESSING (i added some extra columns to make data easier as per the questions)

STEP-3 - DATA ANALYSING (with the help of the pivot table I made table as per the requirement of the store or the answers they want to know)

STEP-4 - In the end, by basic formatting I pasted all the charts in one place and applied some slicers as well so that if I want to see only some specific data it will help me out.

Here are some insights- 
1) December has maximum sales.
2) Women are more likely to buy the products compared to the men.
3) 92% products are delivered.
4) Amazon has maximum contribution.
5) In States, Maharashtra has maximum sales among all.